refact: viewmodel created
